Enable INT8 asymmetric in-flight quantization and channel-wise support
- Implement asymmetric INT8 quantization (RTN) in gguf_utils.
- Add support for channel-wise quantization across INT4/INT8 symmetric and asymmetric modes.
- Update modeling API (Qwen3 MoE) and ops to support group size configuration.
- Enhance safetensors weight finalizer to handle INT8 asymmetric dequantization and 3D weights.
- Add new test cases in auto_tests.py for channel-wise and INT8 asymmetric in-flight quantization.

+++ src/cpp/src/modeling/ops/ops.cpp
@@ -109,7 +109,7 @@ Tensor moe3gemm_fused_compressed(const Tensor& input,
                                  int32_t inter_size,
                                  int32_t num_experts,
                                  int32_t top_k,
-                                 int32_t group_size,
+                                 size_t group_size,
                                  const ov::element::Type& out_type) {
     auto* ctx = input.context();
     auto hidden_f16 = input.to(ov::element::f16);
